Add python packages for SVL

+++ setup/install.sh
@@ -3,13 +3,15 @@ sudo sh -c 'echo "deb http://packages.ros.org/ros/ubuntu $(lsb_release -sc) main
 sudo apt install curl -y
 curl -s https://raw.githubusercontent.com/ros/rosdistro/master/ros.asc | sudo apt-key add -
 sudo apt update
-sudo apt install ros-melodic-desktop-full -y
+if ! sudo apt install ros-melodic-desktop-full -y; then
+    echo "[System] ROS Install Fail"
+    exit 1
+fi
+
 source /opt/ros/melodic/setup.bash
 sudo apt install python-rosdep python-rosinstall python-rosinstall-generator python-wstool build-essential python-rosdep -y
 sudo rosdep init
 rosdep update
-
-echo "ROS Install Sucess"
 
 # System Dependencies of Ubuntu 18.04 / ROS Melodic
 sudo apt-get update -y
@@ -18,25 +20,33 @@ sudo apt install -y python3-pip python3-colcon-common-extensions python3-setupto
 pip3 install -U setuptools
 
 # Eigen Build
-wget https://gitlab.com/libeigen/eigen/-/archive/3.3.7/eigen-3.3.7.tar.gz
-mkdir eigen
-tar --strip-components=1 -xzvf eigen-3.3.7.tar.gz -C eigen
-cd eigen
-mkdir build
-cd build
-cmake ..
-make
-sudo make install
-cd ../..
-rm -rf eigen eigen-3.3.7.tar.gz
-
-sudo rm /usr/lib/cmake/eigen3/*
-sudo cp /usr/local/share/eigen3/cmake/* /usr/lib/cmake/eigen3
+if ! [ -d "/usr/local/share/eigen3"]; then
+    wget https://gitlab.com/libeigen/eigen/-/archive/3.3.7/eigen-3.3.7.tar.gz
+    mkdir eigen
+    tar --strip-components=1 -xzvf eigen-3.3.7.tar.gz -C eigen
+    cd eigen
+    mkdir build
+    cd build
+    cmake ..
+    if ! make; then
+        cd ../..
+        rm -rf eigen eigen-3.3.7.tar.gz
+        echo "[System] Eigen Build Fail"
+        exit 1
+    fi
+    sudo make install
+    cd ../..
+    rm -rf eigen eigen-3.3.7.tar.gz
+    sudo rm /usr/lib/cmake/eigen3/*
+    sudo cp /usr/local/share/eigen3/cmake/* /usr/lib/cmake/eigen3
 
 # Install rosdep
 cd autoware.ai
 rosdep update
-rosdep install -y --from-paths src --ignore-src --rosdistro melodic
+if ! rosdep install -y --from-paths src --ignore-src --rosdistro melodic; then
+    echo "[System] rosdep install Fail"
+    exit 1
+fi
 cd ..
 
 # Resolve OpenCV version issue
@@ -50,24 +60,41 @@ fi
 # Autoware Build
 cd autoware.ai
 if [ -d "/usr/local/cuda" ]; then
-    AUTOWARE_COMPILE_WITH_CUDA=1 colcon build --cmake-args -DCMAKE_BUILD_TYPE=Release
+    if ! AUTOWARE_COMPILE_WITH_CUDA=1 colcon build --cmake-args -DCMAKE_BUILD_TYPE=Release; then
+        echo "[System] Autoware Build Fail"
+        exit 1
+    fi
 else
-    colcon build --cmake-args -DCMAKE_BUILD_TYPE=Release
+    if ! colcon build --cmake-args -DCMAKE_BUILD_TYPE=Release; then
+        echo "[System] Autoware Build Fail"
+        exit 1
+    fi
 fi
 cd ..
 ln -s $(pwd)/autoware.ai ~/autoware.ai
 source ~/autoware.ai/install/setup.bash
-
-echo "Autoware Build Success"
 
 # Build spiraline_ws
 sudo apt-get install ros-melodic-ackermann-msgs ros-melodic-serial ros-melodic-veldoyne ros-melodic-velodyne-driver -y
 cd spiraline_ws/src
 catkin_init_workspace
 cd ..
-catkin_make
+if ! catkin_make; then
+    echo "[System] spiraline_ws Build Fail"
+    exit 1
+fi
 cd ..
 ln -s $(pwd)/spiraline_ws ~/spiraline_ws
 source ~/spiraline_ws/devel/setup.bash
 
-echo "spiraline_ws Build Success"
+# Build other files
+sudo apt-get install -y ros-melodic-rosbridge-server net-tools
+
+# Python packages
+pip3 install PyYAML rospkg matplotlib opencv-python
+cd setup/svl
+python3 -m pip install -r requirements.txt --user .
+cd ../..
+
+# Allow Firewall
+sudo ufw allow 9090
